refactor(panels): unify panel-specific modules under core/ directories
Move align-core, image-core, and audio-core into Align/core, Image/core,
and Audio/core to match the ThreeD panel layout and simplify import paths.
