+++ src/features/panels/Audio/core/waveformBuffer.ts
