+++ src/features/panels/Image/core/imageWorkerProtocol.ts
