+++ src/features/panels/Image/core/messageFrameAdapter.test.ts
