+++ src/features/panels/Image/core/messageFrameAdapter.ts
